Bugfixes of simd functions.

+++ src/vectors/bf16_simd.hpp
@@ -28,30 +28,30 @@ namespace SIMD_NAMESPACE
 {
 #if SUPPORTS_AVX
 
-		static inline __m256 bfloat16_to_float(__m128i x) noexcept
-		{
+	static inline __m256 bfloat16_to_float(__m128i x) noexcept
+	{
 #if SUPPORTS_AVX2
-			__m256i tmp = _mm256_cvtepu16_epi32(x); // extend 16 bits with zeros to 32 bits
-			tmp = _mm256_slli_epi32(tmp, 16); // shift left by 16 bits while shifting in zeros
+		__m256i tmp = _mm256_cvtepu16_epi32(x); // extend 16 bits with zeros to 32 bits
+		tmp = _mm256_slli_epi32(tmp, 16); // shift left by 16 bits while shifting in zeros
 #else
-			__m128i tmp_lo = _mm_unpacklo_epi16(_mm_setzero_si128(), x); // pad lower half with zeros
-			__m128i tmp_hi = _mm_unpackhi_epi16(_mm_setzero_si128(), x); // pad upper half with zeros
-			__m256i tmp = _mm256_setr_m128i(tmp_lo, tmp_hi); // combine two halves
+		__m128i tmp_lo = _mm_unpacklo_epi16(_mm_setzero_si128(), x); // pad lower half with zeros
+		__m128i tmp_hi = _mm_unpackhi_epi16(_mm_setzero_si128(), x); // pad upper half with zeros
+		__m256i tmp = _mm256_setr_m128i(tmp_lo, tmp_hi); // combine two halves
 #endif /* SUPPORTS_AVX2 */
-			return _mm256_castsi256_ps(tmp);
-		}
-		static inline __m128i float_to_bfloat16(__m256 x) noexcept
-		{
+		return _mm256_castsi256_ps(tmp);
+	}
+	static inline __m128i float_to_bfloat16(__m256 x) noexcept
+	{
 #if SUPPORTS_AVX2
-			__m256i tmp = _mm256_srli_epi32(_mm256_castps_si256(x), 16); // shift right by 16 bits while shifting in zeros
-			return _mm_packs_epi32(get_low(tmp), get_high(tmp)); // pack 32 bits into 16 bits
+		__m256i tmp = _mm256_srli_epi32(_mm256_castps_si256(x), 16); // shift right by 16 bits while shifting in zeros
+		return _mm_packs_epi32(get_low(tmp), get_high(tmp)); // pack 32 bits into 16 bits
 #else
-			__m128i tmp_lo = _mm_srli_epi32(_mm_castps_si128(get_low(x)), 16); // shift right by 16 bits while shifting in zeros
-			__m128i tmp_hi = _mm_srli_epi32(_mm_castps_si128(get_high(x)), 16); // shift right by 16 bits while shifting in zeros
-			return _mm_packs_epi32(tmp_lo, tmp_hi); // pack 32 bits into 16 bits
+		__m128i tmp_lo = _mm_srli_epi32(_mm_castps_si128(get_low(x)), 16); // shift right by 16 bits while shifting in zeros
+		__m128i tmp_hi = _mm_srli_epi32(_mm_castps_si128(get_high(x)), 16); // shift right by 16 bits while shifting in zeros
+		return _mm_packs_epi32(tmp_lo, tmp_hi); // pack 32 bits into 16 bits
 #endif
 
-		}
+	}
 
 #elif SUPPORTS_SSE2 /* if __AVX__ is not defined */
 
@@ -94,7 +94,7 @@ namespace SIMD_NAMESPACE
 		private:
 			SIMD<float> m_data;
 		public:
-			static constexpr size_t length = SIMD<float>::length;
+			static constexpr int64_t length = SIMD<float>::length;
 
 			SIMD() noexcept
 			{
@@ -135,8 +135,8 @@ namespace SIMD_NAMESPACE
 			{
 				assert(ptr != nullptr);
 #if SUPPORTS_AVX
-					__m128i tmp = _mm_loadu_si128(reinterpret_cast<const __m128i*>(ptr));
-					m_data = bfloat16_to_float(tmp);
+				__m128i tmp = _mm_loadu_si128(reinterpret_cast<const __m128i*>(ptr));
+				m_data = bfloat16_to_float(tmp);
 #elif SUPPORTS_SSE2
 				__m128i tmp = _mm_loadu_si64(ptr);
 				m_data = bfloat16_to_float(tmp);
@@ -167,7 +167,7 @@ namespace SIMD_NAMESPACE
 			{
 				assert(ptr != nullptr);
 #if SUPPORTS_AVX
-					_mm_storeu_si128(reinterpret_cast<__m128i*>(ptr), float_to_bfloat16(m_data));
+				_mm_storeu_si128(reinterpret_cast<__m128i*>(ptr), float_to_bfloat16(m_data));
 #elif SUPPORTS_SSE2
 				_mm_storeu_si64(reinterpret_cast<__m128i*>(ptr), float_to_bfloat16(m_data));
 #else
@@ -183,7 +183,7 @@ namespace SIMD_NAMESPACE
 				assert(ptr != nullptr);
 				assert(num <= length);
 #if SUPPORTS_AVX
-					partial_store(float_to_bfloat16(m_data), ptr, num * sizeof(avocado::backend::bfloat16));
+				partial_store(float_to_bfloat16(m_data), ptr, num * sizeof(avocado::backend::bfloat16));
 #elif SUPPORTS_SSE2
 				partial_store(float_to_bfloat16(m_data), ptr, num * sizeof(avocado::backend::bfloat16));
 #else
@@ -204,137 +204,6 @@ namespace SIMD_NAMESPACE
 			}
 	};
 
-/* Half-float vector arithmetics */
-//		static inline SIMD<bfloat16> operator+(SIMD<bfloat16> lhs, SIMD<bfloat16> rhs) noexcept
-//		{
-//			return static_cast<SIMD<float>>(lhs) + static_cast<SIMD<float>>(rhs);
-//		}
-//		static inline SIMD<bfloat16> operator+(SIMD<bfloat16> lhs, float rhs) noexcept
-//		{
-//			return lhs + SIMD<bfloat16>(rhs);
-//		}
-//		static inline SIMD<bfloat16> operator+(float lhs, SIMD<bfloat16> rhs) noexcept
-//		{
-//			return SIMD<bfloat16>(lhs) + rhs;
-//		}
-//		static inline SIMD<bfloat16>& operator+=(SIMD<bfloat16> &lhs, SIMD<bfloat16> rhs) noexcept
-//		{
-//			lhs = lhs + rhs;
-//			return lhs;
-//		}
-//
-//		static inline SIMD<bfloat16> operator-(SIMD<bfloat16> lhs, SIMD<bfloat16> rhs) noexcept
-//		{
-//			return static_cast<SIMD<float>>(lhs) - static_cast<SIMD<float>>(rhs);
-//		}
-//		static inline SIMD<bfloat16> operator-(SIMD<bfloat16> lhs, float rhs) noexcept
-//		{
-//			return lhs - SIMD<bfloat16>(rhs);
-//		}
-//		static inline SIMD<bfloat16> operator-(float lhs, SIMD<bfloat16> rhs) noexcept
-//		{
-//			return SIMD<bfloat16>(lhs) - rhs;
-//		}
-//		static inline SIMD<bfloat16>& operator-=(SIMD<bfloat16> &lhs, SIMD<bfloat16> rhs) noexcept
-//		{
-//			lhs = lhs - rhs;
-//			return lhs;
-//		}
-//		static inline SIMD<bfloat16> operator-(SIMD<bfloat16> x) noexcept
-//		{
-//			return -static_cast<SIMD<float>>(x);
-//		}
-//
-//		static inline SIMD<bfloat16> operator*(SIMD<bfloat16> lhs, SIMD<bfloat16> rhs) noexcept
-//		{
-//			return static_cast<SIMD<float>>(lhs) * static_cast<SIMD<float>>(rhs);
-//		}
-//		static inline SIMD<bfloat16> operator*(SIMD<bfloat16> lhs, float rhs) noexcept
-//		{
-//			return lhs * SIMD<bfloat16>(rhs);
-//		}
-//		static inline SIMD<bfloat16> operator*(float lhs, SIMD<bfloat16> rhs) noexcept
-//		{
-//			return SIMD<bfloat16>(lhs) * rhs;
-//		}
-//		static inline SIMD<bfloat16>& operator*=(SIMD<bfloat16> &lhs, SIMD<bfloat16> rhs) noexcept
-//		{
-//			lhs = lhs * rhs;
-//			return lhs;
-//		}
-//
-//		static inline SIMD<bfloat16> operator/(SIMD<bfloat16> lhs, SIMD<bfloat16> rhs) noexcept
-//		{
-//			return static_cast<SIMD<float>>(lhs) / static_cast<SIMD<float>>(rhs);
-//		}
-//		static inline SIMD<bfloat16> operator/(SIMD<bfloat16> lhs, float rhs) noexcept
-//		{
-//			return lhs / SIMD<bfloat16>(rhs);
-//		}
-//		static inline SIMD<bfloat16> operator/(float lhs, SIMD<bfloat16> rhs) noexcept
-//		{
-//			return SIMD<bfloat16>(lhs) / rhs;
-//		}
-//		static inline SIMD<bfloat16>& operator/=(SIMD<bfloat16> &lhs, SIMD<bfloat16> rhs) noexcept
-//		{
-//			lhs = lhs / rhs;
-//			return lhs;
-//		}
-//
-//		/* Float vector functions */
-//		static inline SIMD<bfloat16> max(SIMD<bfloat16> lhs, SIMD<bfloat16> rhs) noexcept
-//		{
-//			return max(static_cast<SIMD<float>>(lhs), static_cast<SIMD<float>>(rhs));
-//		}
-//		static inline SIMD<bfloat16> min(SIMD<bfloat16> lhs, SIMD<bfloat16> rhs) noexcept
-//		{
-//			return min(static_cast<SIMD<float>>(lhs), static_cast<SIMD<float>>(rhs));
-//		}
-//		static inline SIMD<bfloat16> abs(SIMD<bfloat16> x) noexcept
-//		{
-//			return abs(static_cast<SIMD<float>>(x));
-//		}
-//		static inline SIMD<bfloat16> sqrt(SIMD<bfloat16> x) noexcept
-//		{
-//			return sqrt(static_cast<SIMD<float>>(x));
-//		}
-//		static inline SIMD<bfloat16> square(SIMD<bfloat16> x) noexcept
-//		{
-//			return x * x;
-//		}
-//		static inline SIMD<bfloat16> approx_recipr(SIMD<bfloat16> x) noexcept
-//		{
-//			return approx_recipr(static_cast<SIMD<float>>(x));
-//		}
-//		static inline SIMD<bfloat16> approx_rsqrt(SIMD<bfloat16> x) noexcept
-//		{
-//			return approx_rsqrt(static_cast<SIMD<float>>(x));
-//		}
-//		static inline SIMD<bfloat16> sgn(SIMD<bfloat16> x) noexcept
-//		{
-//			return sgn(static_cast<SIMD<float>>(x));
-//		}
-//
-//		static inline SIMD<bfloat16> exp(SIMD<bfloat16> x) noexcept
-//		{
-//			return exp(static_cast<SIMD<float>>(x));
-//		}
-//		static inline SIMD<bfloat16> log(SIMD<bfloat16> x) noexcept
-//		{
-//			return log(static_cast<SIMD<float>>(x));
-//		}
-//		static inline SIMD<bfloat16> tanh(SIMD<bfloat16> x) noexcept
-//		{
-//			return tanh(static_cast<SIMD<float>>(x));
-//		}
-//		static inline SIMD<bfloat16> expm1(SIMD<bfloat16> x) noexcept
-//		{
-//			return expm1(static_cast<SIMD<float>>(x));
-//		}
-//		static inline SIMD<bfloat16> log1p(SIMD<bfloat16> x) noexcept
-//		{
-//			return log1p(static_cast<SIMD<float>>(x));
-//		}
 }
 
 #endif /* VECTORS_BF16_SIMD_HPP_ */
